# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Multi-Business Platform API",
    "description": "A comprehensive backend API supporting hotels, ecommerce, weddings, automobiles, and business websites",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@yourdomain.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Booking": {
        "type": "object",
        "required": ["business", "bookingType", "customer"],
        "properties": {
          "business": {
            "type": "string",
            "description": "Business ID"
          },
          "bookingType": {
            "type": "string",
            "enum": ["hotel", "wedding", "service"]
          },
          "customer": {
            "type": "string",
            "description": "Customer ID"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/bookings": {
      "get": {
        "summary": "Get user bookings",
        "tags": ["Bookings"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of bookings"
          },
          "401": {
            "description": "Not authorized"
          }
        }
      },
      "post": {
        "summary": "Create booking",
        "tags": ["Bookings"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Booking"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created successfully"
          },
          "401": {
            "description": "Not authorized"
          }
        }
      }
    }
  },
...
}